telegram_bot/requirements.txt
python-telegram-bot==21.5
APScheduler==3.10.4
neo4j>=5.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
